refactor(infra): merge the deployment micro-modules by subject
deployment/ held six files that each answered part of one question. They
now sit with the question they belong to.

environment.ts absorbs oidc.ts and stage.ts, so the module that already
owned SST_APP_NAME and the AWS region owns the stage grammar and the
OIDC issuer rules too; stack/deploy.ts drops from three dynamic imports
to one. runtimeBoundaryPolicyArn moves here from bootstrap/, which
removes the last deployment-to-bootstrap import in production code and
tightens the stage it accepts to the grammar every other generated name
uses.

key-policy.ts is storable-keys.ts plus validate-environment.ts. Both
answered 'may this key cross the boundary between a machine and a
stage's shared store', every caller had to ask both, and a key
classified by one file alone is how the two drift apart.

readStoredStageConfig replaces the StageConfigStore class, which was a
constructor and one method around a single call.
